% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box2cox.R
\name{box2cox}
\alias{box2cox}
\title{box2cox Transform Function}
\usage{
box2cox(x, lambda)
}
\arguments{
\item{x}{is the input of this function. x must be positive.}

\item{lambda}{parameter}
}
\value{
y is the output of this function.
}
\description{
box2cox Transform Function
}
