% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box2cox_2parameter.R
\name{box2cox_2parameter}
\alias{box2cox_2parameter}
\title{Two parameter box2cox Transform Function}
\usage{
box2cox_2parameter(x, lambda1, lambda2)
}
\arguments{
\item{x}{is the input of this function. x must be positive.}

\item{lambda1}{parameter}

\item{lambda12}{parameter}
}
\value{
y is the output of this function.
}
\description{
Two parameter box2cox Transform Function
}
