% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box2cox_reverse.R
\name{box2cox_reverse}
\alias{box2cox_reverse}
\title{Reverse box2cox Transform Function}
\usage{
box2cox_reverse(x, lambda)
}
\arguments{
\item{x}{is the input of this function.}

\item{lambda}{parameter}
}
\value{
y is the output of this function.
}
\description{
Reverse box2cox Transform Function
}
